lxd/storage: make (d *powerflex) discover() method to be a generic helper
Move (d *powerflex) discover() out of PowerFlex driver to NVMe connector
code and make it a generic helper.

No functional changes intended.
