+++ lxd/storage/connectors/connector_nvme.go
@@ -2,6 +2,7 @@ package connectors
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"io/fs"
@@ -9,9 +10,11 @@ import (
 	"path/filepath"
 	"slices"
 	"strings"
+	"time"
 
 	"github.com/canonical/lxd/lxd/util"
 	"github.com/canonical/lxd/shared"
+	"github.com/canonical/lxd/shared/logger"
 	"github.com/canonical/lxd/shared/revert"
 )
 
@@ -19,6 +22,14 @@ var _ Connector = &connectorNVMe{}
 
 type connectorNVMe struct {
 	common
+}
+
+type nvmeDiscoveryLogRecord struct {
+	SubNQN string `json:"subnqn"`
+}
+
+type nvmeDiscoveryLog struct {
+	Records []nvmeDiscoveryLogRecord `json:"records"`
 }
 
 // Type returns the type of the connector.
@@ -219,4 +230,54 @@ func (c *connectorNVMe) findSession(targetQN string) (*session, error) {
 	}
 
 	return session, nil
+}
+
+// Discover returns the targets found on the first reachable targetAddr.
+func Discover(ctx context.Context, c Connector, serverUUID string, targetAddresses ...string) ([]nvmeDiscoveryLogRecord, error) {
+	if c.Type() != TypeNVME {
+		return nil, fmt.Errorf("Discover() helper can only be used with NVMe connector type")
+	}
+
+	hostNQN, err := c.QualifiedName()
+	if err != nil {
+		return nil, err
+	}
+
+	// Set a deadline for the overall discovery.
+	ctx, cancel := context.WithTimeout(ctx, 30*time.Second)
+	defer cancel()
+
+	var discoveryLog nvmeDiscoveryLog
+	for _, targetAddr := range targetAddresses {
+		stdout, err := shared.RunCommandContext(ctx, "nvme", "discover", "--transport", "tcp", "--traddr", targetAddr, "--hostnqn", hostNQN, "--hostid", serverUUID, "--output-format", "json")
+		if err != nil {
+			// Exit code 110 is returned if the target address cannot be reached.
+			logger.Warn("Failed connecting to discovery target", logger.Ctx{"target_address": targetAddr, "err": err})
+			continue
+		}
+
+		// In case no discovery log entries can be fetched the nvme command doesn't return JSON formatted text.
+		if strings.Trim(stdout, "\n") == "No discovery log entries to fetch." {
+			logger.Warn("Failed to find discovery log entries", logger.Ctx{"target_address": targetAddr, "err": err})
+			continue
+		}
+
+		// Try to unmarshal the returned log entries.
+		err = json.Unmarshal([]byte(stdout), &discoveryLog)
+		if err != nil {
+			// Don't just log this error.
+			// Something is clearly wrong with the returned output.
+			return nil, fmt.Errorf("Failed to unmarshal the returned discovery log entries from %q", targetAddr)
+		}
+
+		// Unmarshalling the response from the discovery succeeded, break the loop.
+		break
+	}
+
+	// In case none of the target addresses returned any log records also return an error.
+	if len(discoveryLog.Records) == 0 {
+		return nil, errors.New("Failed to fetch a discovery log record from any of the target addresses")
+	}
+
+	return discoveryLog.Records, nil
 }
